refactor: use react-router instead of next/router
