+++ client/next.config.js
@@ -6,4 +6,12 @@ module.exports = {
     };
     return config;
   },
+  // async rewrites() {
+  //   return [
+  //     {
+  //       source: '/:any*',
+  //       destination: '/',
+  //     },
+  //   ];
+  // },
 };
