+++ client/package.json
@@ -15,7 +15,8 @@
     "react-burger-menu": "^3.0.3",
     "react-dom": "^17.0.1",
     "react-hamburger-menu": "^1.2.1",
-    "react-icons": "^4.1.0"
+    "react-icons": "^4.1.0",
+    "react-router-dom": "^5.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.12.10",
